requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
browser-cookie3==0.19.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
curl_cffi==0.6.0b9
discord.py==2.3.2
docstring-inheritance==2.1.2
duckduckgo_search==4.3.1
Flask==3.0.2
Flask-Cors==4.0.0
frozenlist==1.4.1
g4f==0.2.1.9
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Js2Py==0.74
loguru==0.7.2
lxml==5.1.0
lz4==4.3.3
MarkupSafe==2.1.5
multidict==6.0.5
nest-asyncio==1.6.0
numpy==1.26.4
pillow==10.2.0
platformdirs==4.2.0
pycparser==2.21
pycryptodome==3.20.0
pycryptodomex==3.20.0
PyExecJS==1.5.1
pyjsparser==2.7.1
python-dotenv==1.0.0
regex==2023.12.25
requests==2.31.0
six==1.16.0
tiktoken==0.6.0
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.2.0
waitress==3.0.0
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.4